DTOs added

+++ Application/Services/ProductService.cs
@@ -1,5 +1,7 @@
+using Application.Models;
 using Application.Repositories.Interfaces;
 using Application.Services.Interfaces;
+using AutoMapper;
 using Domain.Entities;
 
 namespace Application.Services
@@ -7,33 +9,76 @@ namespace Application.Services
     public class ProductService : IProductService
     {
         private readonly IProductRepository _repo;
-        public ProductService(IProductRepository repo)
+        private readonly IMapper _mapper;
+        
+        public ProductService(IProductRepository repo, IMapper mapper)
         {
+            _mapper = mapper;
             _repo = repo;
         }
 
-        public async Task<Product?> GetProduct (int id) 
+        public async Task<ProductDTO?> GetProduct (int id) 
         {
-           return await _repo.GetById(id);   
+           var product = await _repo.GetById(id);
+           if (product == null) {
+            return null;
+           }
+           return _mapper.Map<Product,ProductDTO>(product);
         }
 
-        public async Task<IEnumerable<Product>> GetAllProducts ()
+        public async Task<IEnumerable<ProductDTO>> GetAllProducts()
         {
-            return await _repo.GetAll();
+            var products = await _repo.GetAll();
+            var productDTO = _mapper.Map<IEnumerable<Product>,IEnumerable<ProductDTO>>(products);
+
+            return productDTO;
         }
 
-        public async Task<bool> AddProduct (Product product)
+
+        public async Task<bool> AddProduct (ProductDTO productDTO)
         {
+            var product = _mapper.Map<ProductDTO, Product>(productDTO);
             return await _repo.Add(product);
         }
 
-        public async Task<bool> UpdateProduct (Product product)
+        public async Task<bool> UpdateProduct (int id, UpsertProductDTO productDTO)
         {
-            return await _repo.Update(product);
+            var updatedProd = await _repo.GetById(id);
+            if (updatedProd == null) 
+            {
+                return false;
+            }
+
+            updatedProd = _mapper.Map(productDTO, updatedProd);
+
+            return await _repo.Update(updatedProd);
         }
 
-        public async Task<bool> RemoveProduct(Product product)
+        public async Task<ProductDTO?> PatchProduct(int id, PatchProductDTO patchedProductDTO)
         {
+            var patchedProd = await _repo.GetById(id);
+            if (patchedProd == null) 
+            {
+                return null;
+            }
+
+            patchedProd = _mapper.Map(patchedProductDTO, patchedProd);
+
+            await _repo.Update(patchedProd);
+
+            var productDTO = _mapper.Map<Product, ProductDTO>(patchedProd);
+
+            return productDTO;
+        }
+
+        public async Task<bool> RemoveProduct(int id)
+        {
+            var product = await _repo.GetById(id);
+
+            if (product == null) {
+                return false;
+            }
+
             return await _repo.Remove(product);
         }
